Some cleanup / moving objects to proper folders / domains
