+++ m3/Assets/Scripts/Source/Gameplay/Controllers/GameController.cs
@@ -2,9 +2,10 @@
 using System.Threading.Tasks;
 using OldBard.Match3.Gameplay.Views.UI;
 using OldBard.Services.Match3.Audio;
-using OldBard.Services.Match3.Factories;
+using OldBard.Services.Match3.Config.Data;
 using OldBard.Services.Match3.Grid;
-using OldBard.Services.Match3.Grid.Data;
+using OldBard.Services.Match3.Grid.Tiles.Data;
+using OldBard.Services.Match3.Grid.Tiles.Factories;
 using OldBard.Services.Match3.Grid.View;
 using UnityEngine;
 using UnityEngine.Pool;
