+++ m3/Assets/Scripts/Source/Gameplay/Views/UI/GameUIController.cs
@@ -1,6 +1,6 @@
 ﻿using System;
 using System.Threading.Tasks;
-using OldBard.Services.Match3.Grid.Data;
+using OldBard.Services.Match3.Config.Data;
 using OldBard.Services.Match3.Grid.View;
 using TMPro;
 using UnityEngine;
