+++ m3/Assets/Scripts/Source/Meta/com.oldbard.match3.meta.asmdef
@@ -2,10 +2,8 @@
   "name": "com.oldbard.match3.meta",
   "rootNamespace": "OldBard.Match3",
   "references": [
-	"com.oldbard.match3.gameplay",
 	"com.oldbard.services.match3.grid",
-	"Unity.TextMeshPro",
-	"Unity"
+	"Unity.TextMeshPro"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
