+++ m3/Packages/com.oldbard.services.match3/Runtime/Grid/GridMatchesHelper.cs
@@ -1,4 +1,5 @@
 ﻿using System.Collections.Generic;
+using OldBard.Services.Match3.Grid.Tiles.Data;
 using UnityEngine;
 using UnityEngine.Pool;
 using Random = System.Random;
