+++ m3/Packages/com.oldbard.services.match3/Runtime/Grid/GridService.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections.Generic;
-using OldBard.Services.Match3.Factories;
+using OldBard.Services.Match3.Grid.Tiles.Data;
+using OldBard.Services.Match3.Grid.Tiles.Factories;
 using UnityEngine;
 using UnityEngine.Pool;
 using Random = System.Random;
