+++ m3/Packages/com.oldbard.services.match3/Runtime/Grid/IGridService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.Collections.Generic;
+using OldBard.Services.Match3.Grid.Tiles.Data;
 using UnityEngine;
 
 namespace OldBard.Services.Match3.Grid
