+++ m3/Packages/com.oldbard.services.match3/Runtime/Utils/CameraAspectRatioPositionSetter.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace OldBard.Services.Utils
+namespace OldBard.Services.Match3.Utils
 {
 	[RequireComponent(typeof(Camera))]
 	public class CameraAspectRatioPositionSetter : MonoBehaviour
