+++ m3/Packages/com.oldbard.services.match3/Runtime/Utils/OrientationAnchorUpdater.cs
@@ -1,6 +1,6 @@
 ﻿using UnityEngine;
 
-namespace OldBard.Services.Utils
+namespace OldBard.Services.Match3.Utils
 {
     /// <summary>
     /// Class that changes the anchors based on the orientation
