+++ m3/Packages/com.oldbard.services.match3/Runtime/Utils/OrientationAnchorUpdaterCustomEditor.cs
@@ -1,7 +1,7 @@
 ﻿using UnityEditor;
 using UnityEngine;
 
-namespace OldBard.Services.Utils
+namespace OldBard.Services.Match3.Utils
 {
 	[CustomEditor(typeof(OrientationAnchorUpdater))]
 	public class OrientationAnchorUpdaterCustomEditor : Editor
