+++ m3/Packages/com.oldbard.services.match3/Runtime/Utils/com.oldbard.services.match3.utils.asmdef
@@ -1,9 +1,8 @@
 {
-  "name": "com.oldbard.services.utils",
-  "rootNamespace": "OldBard.Services.Utils",
+  "name": "com.oldbard.services.match3.utils",
+  "rootNamespace": "OldBard.Services.Match3",
   "references": [
-	"Unity.TextMeshPro",
-	"Unity"
+	"Unity.TextMeshPro"
   ],
   "includePlatforms": [],
   "excludePlatforms": [],
